Fixes for gfortran (source compiles and runs with gfortran version 8 and above)
git-svn-id: https://svn.code.sf.net/p/olam-model/code/trunk@761 2bc74c4e-f164-4307-89b2-97abcff1156f

+++ CMAQ/cgrid_defn.f90
@@ -34,7 +34,8 @@ contains
     implicit none
 
     integer, intent(in) :: mza, mwa
-    integer, parameter  :: cmin = 1.e-30
+
+    real,    parameter  :: cmin = 1.e-30
 
     allocate( cgrid( mza, mwa, nspcsd ) )
     cgrid = cmin
